decimal to float

+++ db/migrate/20191021151120_create_matches.rb
@@ -3,8 +3,8 @@ class CreateMatches < ActiveRecord::Migration[6.0]
     create_table :matches do |t|
       t.integer :user_id
       t.string :place_id
-      t.decimal :place_lat
-      t.decimal :place_lng
+      t.float :place_lat
+      t.float :place_lng
       t.text :place_name
       t.boolean :favorite, default: false
